Fix deploy on OS X

+++ _Misc/dev_scripts/deploy/deploy_osx.sh
@@ -84,7 +84,15 @@ fi
 if [ -f "$SOURCEDIR/bin/_packed/$TarGzArName" ]; then
 	rm "$SOURCEDIR/bin/_packed/$TarGzArName";
 fi
-mv "$DeployDir/$TarGzArName" "$SOURCEDIR/bin/_packed/$TarGzArName"
+
+if [ -f "$DeployDir/$TarGzArName" ]; then
+        mv "$DeployDir/$TarGzArName" "$SOURCEDIR/bin/_packed/$TarGzArName"
+else 
+if [ -f "$DeployDir/rw.$TarGzArName" ]; then
+        mv "$DeployDir/rw.$TarGzArName" "$SOURCEDIR/bin/_packed/$TarGzArName"
+else
+        echo "Built DMG file not found. Nothing to move!"
+fi
 
 echo ""
 echo ""
